chore(electron): keep package internal, restore _electron on playwright
- Mark @playwright/electron private and remove from workspace.js so it is no longer published.
- Move the Electron client implementation back into playwright-core (src/electron/) and expose it as `playwright._electron` from inprocess.ts.
- Keep class-electron.md / class-electronapplication.md in docs/src/electron-api/, merged into core type generation; park class-electronfixtures.md inside the playwright-electron package.
- Update installation tests to import `_electron` from `@playwright/test`.
