+++ packages/playwright-core/index.mjs
@@ -24,4 +24,5 @@ export const devices = playwright.devices;
 export const errors = playwright.errors;
 export const request = playwright.request;
 export const _android = playwright._android;
+export const _electron = playwright._electron;
 export default playwright;
