+++ packages/playwright-core/src/electron/DEPS.list
@@ -1,0 +1,4 @@
+[*]
+@isomorphic/**
+@utils/**
+node_modules/debug
