+++ packages/playwright-core/src/inprocess.ts
@@ -17,6 +17,7 @@
 import { nodePlatform } from '@utils/nodePlatform';
 import { AndroidServerLauncherImpl } from './androidServerImpl';
 import { BrowserServerLauncherImpl } from './browserServerImpl';
+import { Electron } from './electron/electron';
 import { DispatcherConnection, PlaywrightDispatcher, RootDispatcher, createPlaywright } from './server';
 import { Connection } from './client/connection';
 import { packageRoot } from './package';
@@ -43,6 +44,7 @@ export function createInProcessPlaywright(): PlaywrightAPI {
   playwrightAPI.firefox._serverLauncher = new BrowserServerLauncherImpl('firefox');
   playwrightAPI.webkit._serverLauncher = new BrowserServerLauncherImpl('webkit');
   playwrightAPI._android._serverLauncher = new AndroidServerLauncherImpl();
+  (playwrightAPI as any)._electron = new Electron(playwrightAPI);
 
   // Switch to async dispatch after we got Playwright object.
   dispatcherConnection.onmessage = message => setImmediate(() => clientConnection.dispatch(message));
