+++ packages/playwright-electron/src/DEPS.list
@@ -1,5 +1,2 @@
 [*]
-@isomorphic/**
-@utils/**
 node_modules/playwright
-node_modules/debug
