--- utils/generate_types/overrides-electron.d.ts
@@ -1,33 +1,0 @@
-/**
- * Copyright (c) Microsoft Corporation.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-import type { BrowserContext, ConsoleMessage, JSHandle, Page } from 'playwright-core';
-import type { ChildProcess } from 'child_process';
-
-export * from 'playwright-core';
-
-//@ts-ignore this will be any if electron is not installed
-type ElectronType = typeof import('electron');
-
-export interface ElectronApplication {
-  evaluate: JSHandle<ElectronType>['evaluate'];
-  evaluateHandle: JSHandle<ElectronType>['evaluateHandle'];
-}
-
-export const electron: Electron;
-
-// This is required to not export everything by default. See https://github.com/Microsoft/TypeScript/issues/19545#issuecomment-340490459
-export {};
